requirements.txt
fastmcp
pydantic
requests
python-dotenv
uvicorn
